Hook up Kernel.exec to CRuby logic
Follow-up work for jruby/jruby#9609, also connecting Kernel.exec
to the updated CRuby process logic. This includes a large
refactoring that moves the old logic into Helpers and deprecates
the old entry points for that logic. The old logic will only be
used when one of the following is true:

* Running on Windows, for which we have minimal native process
  support and no direct support in the CRuby logic
* Running with NATIVE_EXEC disabled
* Running without native POSIX support

Additional utility methods needed to support the old logic were
also refactored and moved to Helpers.
